Refactored initialization
