+++ src/init/handlers.ts
@@ -4,7 +4,7 @@ import { mousePosition } from 'common/dom';
 import store from 'store';
 import { showMenu, makeSelection } from 'store/actions';
 
-import { Highlighter, TextSelector } from './index';
+import { Highlighter, TextSelector } from '../core/index';
 import {hideMenu} from 'store/widgets/actions';
 import {noSelection} from 'store/textSelector/actions';
