+++ src/store/store.ts
@@ -3,6 +3,7 @@ import thunk from 'redux-thunk';
 import rootReducer from './reducer';
 import promise from 'redux-promise';
 import { createLogger } from 'redux-logger';
+import {  setAxiosConfig } from 'redux-json-api';
 
 // TS override
 declare global {
@@ -15,5 +16,9 @@ const store = createStore(
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
   applyMiddleware(thunk, promise, logger),
 );
+
+store.dispatch(setAxiosConfig({
+  baseURL: 'http://localhost:8000/api',
+}));
 
 export default store;
